Preview feature moved to android specific image component
