--- packages/image/__tests__/android/activate-preview-image.android.test.js
@@ -1,15 +1,0 @@
-/* eslint-env jest */
-
-import tests from "../activate-preview-image.test";
-
-jest.mock("react-native", () => {
-  const reactNative = require.requireActual("react-native");
-  reactNative.Platform.OS = "android";
-  jest
-    .spyOn(reactNative.Platform, "select")
-    .mockImplementation(obj => obj.android || obj.default);
-  return reactNative;
-});
-describe("ActivatePreviewImage test on android", () => {
-  tests();
-});
