--- packages/image/__tests__/android/image-with-preview.android.test.js
@@ -1,16 +1,0 @@
-/* eslint-env jest */
-
-import imageTests from "../image-with-preview.native.test";
-
-jest.mock("react-native", () => {
-  const reactNative = require.requireActual("react-native");
-  reactNative.Platform.OS = "android";
-  jest
-    .spyOn(reactNative.Platform, "select")
-    .mockImplementation(obj => obj.android || obj.default);
-  return reactNative;
-});
-
-describe("ImageWithPreview test on android", () => {
-  imageTests();
-});
