--- packages/image/__tests__/android/image-without-preview.android.test.js
@@ -1,16 +1,0 @@
-/* eslint-env jest */
-
-import imageTests from "../image-without-preview.native.test";
-
-jest.mock("react-native", () => {
-  const reactNative = require.requireActual("react-native");
-  reactNative.Platform.OS = "android";
-  jest
-    .spyOn(reactNative.Platform, "select")
-    .mockImplementation(obj => obj.android || obj.default);
-  return reactNative;
-});
-
-describe("ImageWithoutPreview test on android", () => {
-  imageTests();
-});
