--- packages/image/__tests__/ios/activate-preview-image.ios.test.js
@@ -1,16 +1,0 @@
-/* eslint-env jest */
-
-import tests from "../activate-preview-image.test";
-
-jest.mock("react-native", () => {
-  const reactNative = require.requireActual("react-native");
-  reactNative.Platform.OS = "ios";
-  jest
-    .spyOn(reactNative.Platform, "select")
-    .mockImplementation(obj => obj.ios || obj.default);
-  return reactNative;
-});
-
-describe("ActivatePreviewImage test on ios", () => {
-  tests();
-});
