--- packages/image/__tests__/ios/image-without-preview.ios.test.js
@@ -1,16 +1,0 @@
-/* eslint-env jest */
-
-import imageTests from "../image-without-preview.native.test";
-
-jest.mock("react-native", () => {
-  const reactNative = require.requireActual("react-native");
-  reactNative.Platform.OS = "ios";
-  jest
-    .spyOn(reactNative.Platform, "select")
-    .mockImplementation(obj => obj.ios || obj.default);
-  return reactNative;
-});
-
-describe("ImageWithoutPreview test on ios", () => {
-  imageTests();
-});
