+++ packages/image/__tests__/ios/jest.config.js
@@ -5,6 +5,7 @@ module.exports = Object.assign(jestConfigurator("image", "ios"), {
     "**/packages/image/**/*.js",
     "!**/packages/image/**/*.web.js",
     "!**/packages/image/**/*stories*",
+    "!**/packages/image/**/image.android.js",
     "!**/packages/image/__tests__/**",
     "!**/packages/image/coverage/**"
   ]
