--- packages/image/__tests__/web/activate-preview-image.web.test.js
@@ -1,7 +1,0 @@
-/* eslint-env jest */
-
-import tests from "../activate-preview-image.test";
-
-describe("ActivatePreviewImage test on web", () => {
-  tests();
-});
