+++ packages/image/__tests__/web/jest.config.js
@@ -4,8 +4,7 @@ module.exports = Object.assign(jestConfigurator("image", "web"), {
   collectCoverageFrom: [
     "**/packages/image/**/*.js",
     "!**/packages/image/**/image.js",
-    "!**/packages/image/**/image-with-preview.js",
-    "!**/packages/image/**/image-without-preview.js",
+    "!**/packages/image/**/image.android.js",
     "!**/packages/image/**/add-missing-protocol.js",
     "!**/packages/image/**/style-native.js",
     "!**/packages/image/**/*stories*",
